Splitting iree_trace_replay_* into its own library.
