+++ iree/tools/BUILD
@@ -297,15 +297,13 @@ cc_binary(
     deps = [
         "//iree/base",
         "//iree/base:tracing",
-        "//iree/base/internal:file_io",
         "//iree/base/internal:file_path",
         "//iree/base/internal:flags",
         "//iree/hal",
         "//iree/hal/drivers",
-        "//iree/modules/hal",
+        "//iree/tools/utils:trace_replay",
         "//iree/tools/utils:yaml_util",
         "//iree/vm",
-        "//iree/vm:bytecode_module",
         "@com_github_yaml_libyaml//:yaml",
     ],
 )
